Major refactoring, readme and license
