--- src/main/java/qupath/ext/demo/parameters/DemoExtension.java
@@ -1,89 +1,0 @@
-package qupath.ext.demo.parameters;
-
-import javafx.beans.property.BooleanProperty;
-import javafx.scene.control.MenuItem;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import qupath.lib.common.Version;
-import qupath.lib.gui.QuPathGUI;
-import qupath.lib.gui.dialogs.Dialogs;
-import qupath.lib.gui.extensions.QuPathExtension;
-import qupath.lib.gui.prefs.PathPrefs;
-
-
-/**
- * This is a demo to provide a template for creating a new QuPath extension.
- * <p>
- * It doesn't do much - it just shows how to add a menu item and a preference.
- * See the code and comments below for more info.
- * <p>
- * <b>Important!</b> For your extension to work in QuPath, you need to make sure the name & package
- * of this class is consistent with the file
- * <pre>
- *     /resources/META-INF/services/qupath.lib.gui.extensions.QuPathExtension
- * </pre>
- */
-public class DemoExtension implements QuPathExtension {
-	
-	private final static Logger logger = LoggerFactory.getLogger(DemoExtension.class);
-
-	/**
-	 * Display name for your extension
-	 */
-	private final static String EXTENSION_NAME = "Image processing parameter demo";
-
-	/**
-	 * Short description, used under 'Extensions > Installed extensions'
-	 */
-	private final static String EXTENSION_DESCRIPTION = "This is a demo to show how image processing parameters impact";
-
-	/**
-	 * QuPath version that the extension is designed to work with.
-	 * This allows QuPath to inform the user if it seems to be incompatible.
-	 */
-	private final static Version EXTENSION_QUPATH_VERSION = Version.parse("v0.4.3");
-
-	/**
-	 * Flag whether the extension is already installed (might not be needed... but we'll do it anyway)
-	 */
-	private boolean isInstalled = false;
-
-	@Override
-	public void installExtension(QuPathGUI qupath) {
-		if (isInstalled) {
-			logger.debug("{} is already installed", getName());
-			return;
-		}
-		isInstalled = true;
-		addMenuItem(qupath);
-	}
-
-	/**
-	 * Demo showing how a new command can be added to a QuPath menu.
-	 * @param qupath
-	 */
-	private void addMenuItem(QuPathGUI qupath) {
-		var menu = qupath.getMenu("Extensions>" + EXTENSION_NAME, true);
-		MenuItem menuItem = new MenuItem("Parameter demo");
-		var command = new ParameterAdjustmentCommand(qupath);
-		menuItem.setOnAction(e -> command.run());
-		menu.getItems().add(menuItem);
-	}
-	
-	
-	@Override
-	public String getName() {
-		return EXTENSION_NAME;
-	}
-
-	@Override
-	public String getDescription() {
-		return EXTENSION_DESCRIPTION;
-	}
-	
-	@Override
-	public Version getQuPathVersion() {
-		return EXTENSION_QUPATH_VERSION;
-	}
-
-}
