+++ src/main/resources/META-INF/services/qupath.lib.gui.extensions.QuPathExtension
@@ -1,1 +1,1 @@
-qupath.ext.demo.parameters.DemoExtension
+qupath.ext.demo.parameters.ParameterDemoExtension
